refactor: 회원 가입시 이미지가 없는 경우도 허용, 삭제/수정시 본인의 정보에만 접근 가능하도록 수정

+++ src/main/java/admin/adminsiteserver/member/member/application/MemberService.java
@@ -2,6 +2,8 @@ package admin.adminsiteserver.member.member.application;
 
 import admin.adminsiteserver.common.aws.infrastructure.S3Uploader;
 import admin.adminsiteserver.common.aws.infrastructure.dto.FilePathDto;
+import admin.adminsiteserver.member.auth.util.LoginUser;
+import admin.adminsiteserver.member.auth.util.dto.LoginUserInfo;
 import admin.adminsiteserver.member.member.application.dto.MemberDto;
 import admin.adminsiteserver.member.member.domain.Member;
 import admin.adminsiteserver.member.member.domain.MemberFilePath;
@@ -9,6 +11,8 @@ import admin.adminsiteserver.member.member.domain.MemberFilePathRepository;
 import admin.adminsiteserver.member.member.domain.MemberRepository;
 import admin.adminsiteserver.member.member.exception.AlreadyExistUserIDException;
 import admin.adminsiteserver.member.member.exception.NotExistMemberException;
+import admin.adminsiteserver.member.member.exception.NotPossibleDeleteOtherException;
+import admin.adminsiteserver.member.member.exception.NotPossibleUpdateOtherException;
 import admin.adminsiteserver.member.member.ui.dto.SignUpRequest;
 import admin.adminsiteserver.member.member.ui.dto.UpdateMemberRequest;
 import lombok.RequiredArgsConstructor;
@@ -16,6 +20,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -35,17 +41,27 @@ public class MemberService {
                 .ifPresent(m -> {
                     throw new AlreadyExistUserIDException();
                 });
-        FilePathDto filePathDto = s3Uploader.upload(signUpRequest.getImage(), MEMBER_IMAGE_PATH);
         Member signupMember = memberRepository.save(member);
 
-        signupMember.addProfileImage(filePathDto.toFilePath(MemberFilePath.class));
-        return MemberDto.of(signupMember, filePathDto);
+        if (signUpRequest.getImage() != null) {
+            FilePathDto filePathDto = s3Uploader.upload(signUpRequest.getImage(), MEMBER_IMAGE_PATH);
+            signupMember.addProfileImage(filePathDto.toFilePath(MemberFilePath.class));
+            return MemberDto.of(signupMember, filePathDto);
+        }
+        return MemberDto.from(signupMember);
     }
 
     @Transactional
-    public void updateMember(UpdateMemberRequest updateMemberRequest, String userId) {
+    public void updateMember(UpdateMemberRequest updateMemberRequest, String userId, LoginUserInfo loginUserInfo) {
         Member member = memberRepository.findByUserId(userId)
                 .orElseThrow(NotExistMemberException::new);
+        if (!Objects.equals(userId, loginUserInfo.getUserId())) {
+            throw new NotPossibleUpdateOtherException();
+        }
+        if(updateMemberRequest.getImage() != null) {
+            FilePathDto filePathDto = s3Uploader.upload(updateMemberRequest.getImage(), MEMBER_IMAGE_PATH);
+            member.addProfileImage(filePathDto.toFilePath(MemberFilePath.class));
+        }
         member.update(updateMemberRequest.getEmail(),
                 updateMemberRequest.getName(),
                 updateMemberRequest.getStudentNumber(),
@@ -62,7 +78,10 @@ public class MemberService {
     }
 
     @Transactional
-    public void deleteMember(String userId) {
+    public void deleteMember(String userId, LoginUserInfo loginUserInfo) {
+        if (!Objects.equals(userId, loginUserInfo.getUserId())) {
+            throw new NotPossibleDeleteOtherException();
+        }
         memberRepository.deleteByUserId(userId);
     }
 }
